ECL-978, implementing remarks

+++ tools/records-downloader-tool/src/test/java/eu/europeana/cloud/util/FolderCompressorTest.java
@@ -9,20 +9,23 @@ import org.junit.Test;
 import org.zeroturnaround.zip.ZipException;
 
 import java.io.*;
+import java.util.Date;
 
-import static junit.framework.TestCase.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
 public class FolderCompressorTest implements TestConstantsHelper {
     String folderPath;
+    String zipFolderPath;
 
     @Test(expected = ZipException.class)
     public void shouldThrowZipExceptionWhileCompressEmptyFolder() throws Exception {
         folderPath = FileUtil.createFolder();
         File folder = new File(folderPath);
         assertTrue(folder.isDirectory());
-        FolderCompressor.compress(folderPath);
+        zipFolderPath = FileUtil.createZipFolderPath(new Date());
+        FolderCompressor.compress(folderPath, zipFolderPath);
+        System.out.println(folderPath);
 
     }
 
@@ -33,15 +36,17 @@ public class FolderCompressorTest implements TestConstantsHelper {
         assertTrue(folder.isDirectory());
         InputStream inputStream = IOUtils.toInputStream("some test data for my input stream");
         createFile(inputStream, folderPath + "fileName");
-        String zipFolderPath = FolderCompressor.compress(folderPath);
+        zipFolderPath = FileUtil.createZipFolderPath(new Date());
+        FolderCompressor.compress(folderPath, zipFolderPath);
         assertNotNull(zipFolderPath);
-        assertFalse(folder.exists());
-        FileUtils.forceDelete(new File(zipFolderPath));
+
     }
 
     @After
     public void removeTempFolder() throws IOException {
         FileUtils.deleteDirectory(new File(folderPath));
+        if (zipFolderPath != null)
+            FileUtils.forceDelete(new File(zipFolderPath));
     }
